more codeql issues
