+++ OpenWeatherMap.Avalonia.Sample/ViewModels/MainWindowViewModel.cs
@@ -1,13 +1,10 @@
 ﻿using Avalonia.Media.Imaging;
-using Avalonia.Platform;
 using OpenWeatherMap.Avalonia.Sample.Models;
 using OpenWeatherMap.Standard;
 using OpenWeatherMap.Standard.Enums;
 using OpenWeatherMap.Standard.Models;
 using System;
-using System.IO;
 using System.Linq;
-using System.Net.Http;
 using System.Threading.Tasks;
 
 namespace OpenWeatherMap.Avalonia.Sample.ViewModels
@@ -22,7 +19,6 @@ namespace OpenWeatherMap.Avalonia.Sample.ViewModels
         private string windSpeed;
         private string windDirection;
 
-        public string Greeting => "Welcome to Avalonia!";
         public string CurrentWeather
         {
             get => currentWeather;
@@ -111,31 +107,6 @@ namespace OpenWeatherMap.Avalonia.Sample.ViewModels
             });
             Task.WaitAll();
 
-        }
-    }
-
-    public static class ImageHelper
-    {
-        public static Bitmap LoadFromResource(Uri resourceUri)
-        {
-            return new Bitmap(AssetLoader.Open(resourceUri));
-        }
-
-        public static async Task<Bitmap?> LoadFromWeb(Uri url)
-        {
-            using var httpClient = new HttpClient();
-            try
-            {
-                var response = await httpClient.GetAsync(url);
-                response.EnsureSuccessStatusCode();
-                var data = await response.Content.ReadAsByteArrayAsync();
-                return new Bitmap(new MemoryStream(data));
-            }
-            catch (HttpRequestException ex)
-            {
-                Console.WriteLine($"An error occurred while downloading image '{url}' : {ex.Message}");
-                return null;
-            }
         }
     }
 }
