+++ OpenWeatherMap.Avalonia.Sample/ViewModels/ViewModelBase.cs
@@ -1,10 +1,10 @@
 ﻿using System.ComponentModel;
-
+#nullable disable
 namespace OpenWeatherMap.Avalonia.Sample.ViewModels
 {
     public class ViewModelBase : INotifyPropertyChanged
     {
-        public event PropertyChangedEventHandler? PropertyChanged;
+        public event PropertyChangedEventHandler PropertyChanged;
 
         protected void OnPropertyChanged(string propertyName)
         {
@@ -12,3 +12,4 @@ namespace OpenWeatherMap.Avalonia.Sample.ViewModels
         }
     }
 }
+#nullable enable
