+++ OpenWeatherMap.Standard.Core.Test/CityNameTests.cs
@@ -1,13 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-using FakeItEasy;
+﻿using FakeItEasy;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;
 using OpenWeatherMap.Standard.Enums;
 using OpenWeatherMap.Standard.Interfaces;
 using OpenWeatherMap.Standard.Models;
+using System.Threading.Tasks;
 
 namespace OpenWeatherMap.Standard.Core.Test
 {
@@ -19,7 +16,7 @@ namespace OpenWeatherMap.Standard.Core.Test
 
         public CityNameTests()
         {
-            expected = JsonConvert.DeserializeObject<WeatherData>(expectedJson);
+            expected = JsonConvert.DeserializeObject<WeatherData>(expectedJson) ?? new WeatherData();
         }
         [TestMethod()]
         public void TestCloudy()
