+++ OpenWeatherMap.Standard/Models/GeoLocation.cs
@@ -1,12 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace OpenWeatherMap.Standard.Models
+﻿namespace OpenWeatherMap.Standard.Models
 {
 
     public class GeoLocation
     {
+        public GeoLocation()
+        {
+            name = string.Empty;
+            country = string.Empty;
+            state = string.Empty;
+        }
         public string name { get; set; }
         public float lat { get; set; }
         public float lon { get; set; }
